refactor: split team/impl.ts + hub.ts + plugin/registry.ts to ≤200 LOC
Round 2 of LOC splits, 3 largest remaining files split by parallel Sonnet agents:

team/impl.ts (549 → 6 files):
  impl.ts (64) — cmdTeamList + barrel re-exports
  team-helpers.ts (99) — shared state, interfaces, writeMessage
  team-lifecycle.ts (200) — cmdTeamShutdown/Create/Spawn
  team-comms.ts (33) — cmdTeamSend
  team-reincarnation.ts (75) — cmdTeamResume, cmdTeamLives
  team-cleanup-zombies.ts (97) — cmdCleanupZombies, findZombiePanes

hub.ts (489 → 4 files):
  hub.ts (31) — barrel
  hub-config.ts (60) — WorkspaceConfig, loadWorkspaceConfigs
  hub-connection.ts (198) — openWebSocket + lifecycle helpers
  hub-transport.ts (198) — HubTransport class

plugin/registry.ts (483 → 4 files):
  registry.ts (163) — barrel + discoverPackages
  registry-semver.ts (77) — satisfies, formatSdkMismatchError
  registry-helpers.ts (79) — runtimeSdkVersion, hashFile, isDevModeInstall
  registry-invoke.ts (192) — invokePlugin

All existing import paths unchanged (barrel re-exports preserve API).
Tests: 884 pass / 0 fail.
