+++ src/commands/plugins/team/team-helpers.ts
@@ -1,0 +1,99 @@
+import { readFileSync, writeFileSync, existsSync, rmSync, mkdirSync } from "fs";
+import { join } from "path";
+import { homedir } from "os";
+
+// Exported for testing — override with _setDirs
+export let TEAMS_DIR = join(homedir(), ".claude/teams");
+export let TASKS_DIR = join(homedir(), ".claude/tasks");
+
+/** @internal — for tests only */
+export function _setDirs(teams: string, tasks: string) {
+  TEAMS_DIR = teams;
+  TASKS_DIR = tasks;
+}
+
+export interface TeamMember {
+  name: string;
+  agentId?: string;
+  agentType?: string;
+  tmuxPaneId?: string;
+  color?: string;
+  model?: string;
+  backendType?: string;
+}
+
+export interface TeamConfig {
+  name: string;
+  description?: string;
+  members: TeamMember[];
+  createdAt?: number;
+}
+
+export function loadTeam(name: string): TeamConfig | null {
+  const configPath = join(TEAMS_DIR, name, "config.json");
+  if (!existsSync(configPath)) return null;
+  try { return JSON.parse(readFileSync(configPath, "utf-8")); }
+  catch { return null; }
+}
+
+/** Resolve ψ/ directory from cwd — the oracle vault root. */
+export function resolvePsi(): string {
+  const psi = join(process.cwd(), "ψ");
+  if (existsSync(psi)) return psi;
+  try {
+    const real = readFileSync(psi, "utf-8"); // will throw if not exists
+    return real;
+  } catch {
+    return psi; // return default — callers mkdir as needed
+  }
+}
+
+/**
+ * Write a shutdown_request message to a teammate's inbox file.
+ * This is the same protocol Claude Code uses internally via SendMessage.
+ */
+export function writeShutdownRequest(teamName: string, memberName: string, reason: string): void {
+  const inboxPath = join(TEAMS_DIR, teamName, "inboxes", `${memberName}.json`);
+  let messages: any[] = [];
+  if (existsSync(inboxPath)) {
+    try { messages = JSON.parse(readFileSync(inboxPath, "utf-8")); } catch { messages = []; }
+  }
+  const requestId = `shutdown-${Date.now()}@${memberName}`;
+  messages.push({
+    from: "maw-team-shutdown",
+    text: JSON.stringify({ type: "shutdown_request", reason, request_id: requestId }),
+    summary: `Shutdown request: ${reason}`,
+    timestamp: new Date().toISOString(),
+    read: false,
+  });
+  writeFileSync(inboxPath, JSON.stringify(messages, null, 2));
+}
+
+/**
+ * Write a generic message to a teammate's inbox file.
+ * Same protocol as writeShutdownRequest but with type: "message".
+ */
+export function writeMessage(teamName: string, memberName: string, from: string, text: string): void {
+  const inboxPath = join(TEAMS_DIR, teamName, "inboxes", `${memberName}.json`);
+  let messages: any[] = [];
+  if (existsSync(inboxPath)) {
+    try { messages = JSON.parse(readFileSync(inboxPath, "utf-8")); } catch { messages = []; }
+  }
+  messages.push({
+    from,
+    text: JSON.stringify({ type: "message", content: text }),
+    summary: text.slice(0, 80),
+    timestamp: new Date().toISOString(),
+    read: false,
+  });
+  const dir = join(TEAMS_DIR, teamName, "inboxes");
+  if (!existsSync(dir)) mkdirSync(dir, { recursive: true });
+  writeFileSync(inboxPath, JSON.stringify(messages, null, 2));
+}
+
+export function cleanupTeamDir(name: string) {
+  const teamDir = join(TEAMS_DIR, name);
+  const tasksDir = join(TASKS_DIR, name);
+  if (existsSync(teamDir)) { try { rmSync(teamDir, { recursive: true }); } catch {} }
+  if (existsSync(tasksDir)) { try { rmSync(tasksDir, { recursive: true }); } catch {} }
+}
